stm32f4xx: Combine dma1.rs and dma2.rs into dma.rs
The dma1 and dma2 submodules were too small to justify their existence -
therefore it makes more sense to combine their code into a single dma
module.
