+++ boards/stm32f429idiscovery/src/main.rs
@@ -127,11 +127,11 @@ impl
 
 /// Helper function called during bring-up that configures DMA.
 unsafe fn setup_dma(
-    dma: &stm32f429zi::dma::dma2::Dma2,
-    dma_streams: &'static [stm32f429zi::dma::Stream<'static, stm32f429zi::dma::dma2::Dma2>; 8],
-    usart1: &'static stm32f429zi::usart::Usart<stm32f429zi::dma::dma2::Dma2>,
+    dma: &stm32f429zi::dma::Dma2,
+    dma_streams: &'static [stm32f429zi::dma::Stream<'static, stm32f429zi::dma::Dma2>; 8],
+    usart1: &'static stm32f429zi::usart::Usart<stm32f429zi::dma::Dma2>,
 ) {
-    use stm32f429zi::dma::dma2::Dma2Peripheral;
+    use stm32f429zi::dma::Dma2Peripheral;
     use stm32f429zi::usart;
 
     dma.enable_clock();
@@ -263,7 +263,7 @@ unsafe fn setup_peripherals(tim2: &stm32f429zi::tim2::Tim2) {
 unsafe fn get_peripherals() -> (
     &'static mut Stm32f429ziDefaultPeripherals<'static>,
     &'static stm32f429zi::syscfg::Syscfg<'static>,
-    &'static stm32f429zi::dma::dma2::Dma2<'static>,
+    &'static stm32f429zi::dma::Dma2<'static>,
 ) {
     // We use the default HSI 16Mhz clock
     let rcc = static_init!(stm32f429zi::rcc::Rcc, stm32f429zi::rcc::Rcc::new());
@@ -275,14 +275,8 @@ unsafe fn get_peripherals() -> (
         stm32f429zi::exti::Exti,
         stm32f429zi::exti::Exti::new(syscfg)
     );
-    let dma1 = static_init!(
-        stm32f429zi::dma::dma1::Dma1,
-        stm32f429zi::dma::dma1::Dma1::new(rcc)
-    );
-    let dma2 = static_init!(
-        stm32f429zi::dma::dma2::Dma2,
-        stm32f429zi::dma::dma2::Dma2::new(rcc)
-    );
+    let dma1 = static_init!(stm32f429zi::dma::Dma1, stm32f429zi::dma::Dma1::new(rcc));
+    let dma2 = static_init!(stm32f429zi::dma::Dma2, stm32f429zi::dma::Dma2::new(rcc));
     let peripherals = static_init!(
         Stm32f429ziDefaultPeripherals,
         Stm32f429ziDefaultPeripherals::new(rcc, exti, dma1, dma2)
