+++ chips/stm32f429zi/src/interrupt_service.rs
@@ -10,8 +10,8 @@ impl<'a> Stm32f429ziDefaultPeripherals<'a> {
     pub unsafe fn new(
         rcc: &'a crate::rcc::Rcc,
         exti: &'a crate::exti::Exti<'a>,
-        dma1: &'a crate::dma::dma1::Dma1<'a>,
-        dma2: &'a crate::dma::dma2::Dma2<'a>,
+        dma1: &'a crate::dma::Dma1<'a>,
+        dma2: &'a crate::dma::Dma2<'a>,
     ) -> Self {
         Self {
             stm32f4: Stm32f4xxDefaultPeripherals::new(rcc, exti, dma1, dma2),
