+++ chips/stm32f4xx/src/chip.rs
@@ -6,7 +6,7 @@ use kernel::deferred_call;
 use kernel::platform::chip::Chip;
 use kernel::platform::chip::InterruptService;
 
-use crate::dma::{dma1, dma2};
+use crate::dma;
 use crate::nvic;
 
 use crate::deferred_calls::DeferredCallTask;
@@ -19,15 +19,15 @@ pub struct Stm32f4xx<'a, I: InterruptService<DeferredCallTask> + 'a> {
 
 pub struct Stm32f4xxDefaultPeripherals<'a> {
     pub adc1: crate::adc::Adc<'a>,
-    pub dma1_streams: [crate::dma::Stream<'a, dma1::Dma1<'a>>; 8],
-    pub dma2_streams: [crate::dma::Stream<'a, dma2::Dma2<'a>>; 8],
+    pub dma1_streams: [crate::dma::Stream<'a, dma::Dma1<'a>>; 8],
+    pub dma2_streams: [crate::dma::Stream<'a, dma::Dma2<'a>>; 8],
     pub exti: &'a crate::exti::Exti<'a>,
     pub i2c1: crate::i2c::I2C<'a>,
     pub spi3: crate::spi::Spi<'a>,
     pub tim2: crate::tim2::Tim2<'a>,
-    pub usart1: crate::usart::Usart<'a, dma2::Dma2<'a>>,
-    pub usart2: crate::usart::Usart<'a, dma1::Dma1<'a>>,
-    pub usart3: crate::usart::Usart<'a, dma1::Dma1<'a>>,
+    pub usart1: crate::usart::Usart<'a, dma::Dma2<'a>>,
+    pub usart2: crate::usart::Usart<'a, dma::Dma1<'a>>,
+    pub usart3: crate::usart::Usart<'a, dma::Dma1<'a>>,
     pub gpio_ports: crate::gpio::GpioPorts<'a>,
     pub fsmc: crate::fsmc::Fsmc<'a>,
 }
@@ -36,13 +36,13 @@ impl<'a> Stm32f4xxDefaultPeripherals<'a> {
     pub fn new(
         rcc: &'a crate::rcc::Rcc,
         exti: &'a crate::exti::Exti<'a>,
-        dma1: &'a dma1::Dma1<'a>,
-        dma2: &'a dma2::Dma2<'a>,
+        dma1: &'a dma::Dma1<'a>,
+        dma2: &'a dma::Dma2<'a>,
     ) -> Self {
         Self {
             adc1: crate::adc::Adc::new(rcc),
-            dma1_streams: dma1::new_dma1_stream(dma1),
-            dma2_streams: dma2::new_dma2_stream(dma2),
+            dma1_streams: dma::new_dma1_stream(dma1),
+            dma2_streams: dma::new_dma2_stream(dma2),
             exti,
             i2c1: crate::i2c::I2C::new(rcc),
             spi3: crate::spi::Spi::new(
@@ -51,8 +51,8 @@ impl<'a> Stm32f4xxDefaultPeripherals<'a> {
                     crate::rcc::PeripheralClockType::APB1(crate::rcc::PCLK1::SPI3),
                     rcc,
                 )),
-                dma1::Dma1Peripheral::SPI3_TX,
-                dma1::Dma1Peripheral::SPI3_RX,
+                dma::Dma1Peripheral::SPI3_TX,
+                dma::Dma1Peripheral::SPI3_RX,
             ),
             tim2: crate::tim2::Tim2::new(rcc),
             usart1: crate::usart::Usart::new_usart1(rcc),
@@ -80,29 +80,29 @@ impl<'a> InterruptService<DeferredCallTask> for Stm32f4xxDefaultPeripherals<'a> 
     unsafe fn service_interrupt(&self, interrupt: u32) -> bool {
         match interrupt {
             nvic::DMA1_Stream1 => self.dma1_streams
-                [dma1::Dma1Peripheral::USART3_RX.get_stream_idx()]
+                [dma::Dma1Peripheral::USART3_RX.get_stream_idx()]
             .handle_interrupt(),
             nvic::DMA1_Stream2 => {
-                self.dma1_streams[dma1::Dma1Peripheral::SPI3_RX.get_stream_idx()].handle_interrupt()
+                self.dma1_streams[dma::Dma1Peripheral::SPI3_RX.get_stream_idx()].handle_interrupt()
             }
             nvic::DMA1_Stream3 => self.dma1_streams
-                [dma1::Dma1Peripheral::USART3_TX.get_stream_idx()]
+                [dma::Dma1Peripheral::USART3_TX.get_stream_idx()]
             .handle_interrupt(),
             nvic::DMA1_Stream5 => self.dma1_streams
-                [dma1::Dma1Peripheral::USART2_RX.get_stream_idx()]
+                [dma::Dma1Peripheral::USART2_RX.get_stream_idx()]
             .handle_interrupt(),
             nvic::DMA1_Stream6 => self.dma1_streams
-                [dma1::Dma1Peripheral::USART2_TX.get_stream_idx()]
+                [dma::Dma1Peripheral::USART2_TX.get_stream_idx()]
             .handle_interrupt(),
             nvic::DMA1_Stream7 => {
-                self.dma1_streams[dma1::Dma1Peripheral::SPI3_TX.get_stream_idx()].handle_interrupt()
+                self.dma1_streams[dma::Dma1Peripheral::SPI3_TX.get_stream_idx()].handle_interrupt()
             }
 
             nvic::DMA2_Stream5 => self.dma2_streams
-                [dma2::Dma2Peripheral::USART1_RX.get_stream_idx()]
+                [dma::Dma2Peripheral::USART1_RX.get_stream_idx()]
             .handle_interrupt(),
             nvic::DMA2_Stream7 => self.dma2_streams
-                [dma2::Dma2Peripheral::USART1_TX.get_stream_idx()]
+                [dma::Dma2Peripheral::USART1_TX.get_stream_idx()]
             .handle_interrupt(),
 
             nvic::USART1 => self.usart1.handle_interrupt(),
